requirements.txt
requests>=2.25.1
pandas>=2.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
cryptography>=45.0.5